Split native sidebar project list
